Refactor: move into routes, controllers and models
